app/requirements.txt
asgiref==3.8.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.3
djangorestframework==3.15.0
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.1.1
pytest-django==4.8.0
python-decouple==3.8
requests==2.31.0
ruff==0.3.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
